requirements.txt
torch
numpy
torchvision
pillow
tqdm